Working constraint on DTD and XSD.
DTD: No changes necessary beyond what is in the tracked changes in the tutorial repo.
XSD: Procedure works on DITA 1.2 Schema, but not 1.3. I do not know enough about Schemas in general, much less the DITA Schemas for 1.3, to make adjustments quickly; rather than prolong the process to accommodate my learning curve, in light of Eliot's remarks about Schema support in DITA 2.0, I am moving on to the RNG piece.